# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Layer - YouTube Creator-Editor Collaboration Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export type RegisterRequestRole =
  (typeof RegisterRequestRole)[keyof typeof RegisterRequestRole];

export const RegisterRequestRole = {
  creator: "creator",
  editor: "editor",
} as const;

export interface RegisterRequest {
  email: string;
  /** @minLength 6 */
  password: string;
  name: string;
  role: RegisterRequestRole;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  creator: "creator",
  editor: "editor",
} as const;

export interface User {
  id: string;
  email: string;
  name: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface UserListResponse {
  users: User[];
}

export interface CreateVideoRequest {
  title: string;
  description: string;
  tags?: string[];
  videoUrl: string;
  thumbnailUrl?: string;
  creatorId: string;
  fileSize?: number;
  duration?: number;
}

export type VideoStatus = (typeof VideoStatus)[keyof typeof VideoStatus];

export const VideoStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
  uploaded: "uploaded",
} as const;

export interface Video {
  id: string;
  title: string;
  description: string;
  tags?: string[];
  videoUrl: string;
  thumbnailUrl?: string;
  status: VideoStatus;
  creatorId: string;
  editorId: string;
  rejectionFeedback?: string;
  fileSize?: number;
  duration?: number;
  createdAt: string;
  updatedAt: string;
  creator?: User;
  editor?: User;
}

export interface VideoListResponse {
  videos: Video[];
}

export interface RejectVideoRequest {
  feedback: string;
}

export type NotificationType =
  (typeof NotificationType)[keyof typeof NotificationType];

export const NotificationType = {
  video_pending: "video_pending",
  video_approved: "video_approved",
  video_rejected: "video_rejected",
  video_uploaded: "video_uploaded",
} as const;

export interface Notification {
  id: string;
  userId: string;
  title: string;
  message: string;
  type: NotificationType;
  read: boolean;
  videoId?: string;
  createdAt: string;
}

export interface NotificationListResponse {
  notifications: Notification[];
}

export type ListVideosParams = {
  status?: ListVideosStatus;
  creatorId?: string;
};

export type ListVideosStatus =
  (typeof ListVideosStatus)[keyof typeof ListVideosStatus];

export const ListVideosStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
  uploaded: "uploaded",
} as const;