# lib/api-zod/src/generated/types/createVideoRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Layer - YouTube Creator-Editor Collaboration Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateVideoRequest {
  title: string;
  description: string;
  tags?: string[];
  videoUrl: string;
  thumbnailUrl?: string;
  creatorId: string;
  fileSize?: number;
  duration?: number;
}