# lib/api-zod/src/generated/types/video.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Layer - YouTube Creator-Editor Collaboration Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from "./user";
import type { VideoStatus } from "./videoStatus";

export interface Video {
  id: string;
  title: string;
  description: string;
  tags?: string[];
  videoUrl: string;
  thumbnailUrl?: string;
  status: VideoStatus;
  creatorId: string;
  editorId: string;
  rejectionFeedback?: string;
  fileSize?: number;
  duration?: number;
  createdAt: Date;
  updatedAt: Date;
  creator?: User;
  editor?: User;
}